requirements.txt
amqp==1.4.9
anyjson==0.3.3
argh==0.26.2
argproc==1.4
asn1crypto==0.24.0
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
crontab==0.22.4
cryptography==2.4.2
Django==2.1.2
django-celery==3.2.2
django-common-helpers==0.9.2
django-cron==0.5.1
django-crontab==0.7.1
django-filter==2.0.0
djangorestframework==3.9.0
future==0.17.1
get==1.0.3
idna==2.7
kombu==3.0.37
Markdown==3.0.1
mysql==0.0.2
mysqlclient==1.3.13
nose==1.3.7
ply==3.11
post==1.0.2
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
public==2.0.1
pycparser==2.19
PyDispatcher==2.0.5
PyMySQL==0.9.2
python-rest==1.3
pytz==2018.5
PyYAML==3.13
query-string==1.0.2
redis==2.10.3
request==1.0.2
requests==2.20.0
six==1.11.0
task==0.0.0
tool==0.8.0
tornado==5.1.1
urllib3==1.24.1
vega==1.4.0
views==0.3
